Refactor: 막대그래프 막대 클릭 시 동작 책임 위치 변경
- 각각의 동작 구현을 상위 위젯에서 하도록 처리
